fix(company-instructions): empty default content and allow deleting any file
- COMPANY.md now starts empty instead of with boilerplate template
- Delete button now appears for all files including the entry file
  (previously hidden for the entry file, leaving users with no way
  to delete COMPANY.md)
